pyproject.toml
[project]
name = "commonroad-control"
version = "0.0.1"
description = "CommonRoad Control Toolbox"
authors = [{name = "Lukas Schäfer", email = "<lukas.schaefer@tum.de>"}, {name="Tobias Mascetta", email = "<tobias.mascetta@tum.de>"}]
license = "BSD"
readme = "README.md"
keywords= ["autonomous", "automated", "vehicles", "driving", "control"]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX :: Linux",
]

packages = [{include = "commonroad_control"}]

requires-python = "^3.9"
dynamic = ["dependencies"]

# Dependencies
[tool.poetry.dependencies]
python = ">=3.9, <3.12"
numpy = ">=2.0, <3.0"
matplotlib = ">=3.9"
casadi = ">=3.6"
scipy = ">=1.13"
clarabel = ">= 0.9.0"
commonroad-io = ">=2022.2"
cvxpy = "1.6.0"



# planner
# test
[tool.poetry.group.planner]
optional = true

[tool.poetry.group.planner.dependencies]
commonroad-reactive-planner = "2025.1"


# test
[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^6.5.0"

# docs
[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocstrings-python = "^1.8.0"
mkdocs-material = "^9.5.10"





[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
